Refactor documentation links and improve structure
- Updated example links in onboarding documentation to point directly to GitHub.
- Changed figure path in mental models documentation for consistency.
- Enhanced persona paths documentation with updated example links.
- Added :no-index: directive to API reference pages to prevent indexing.
- Introduced cache_config to configuration index for better organization.
- Added predictions index to user guide for improved navigation.
- Renamed deep learning example files for clarity and consistency.
- Added generators section to pipelines index for completeness.
- Updated method descriptions in result classes to reflect key operations.
- Improved memory utility function to ensure float return type.

+++ nirs4all/utils/memory.py
@@ -91,8 +91,8 @@ def get_process_rss_mb() -> float:
         # Field 1 (index 1) is RSS in pages
         rss_pages = int(parts[1])
         import resource
-        page_size = resource.getpagesize()
-        return (rss_pages * page_size) / (1024 * 1024)
+        page_size = resource.getpagesize()  # type: ignore[attr-defined]
+        return float((rss_pages * page_size) / (1024 * 1024))
     except (OSError, IndexError, ValueError):
         pass
